Library: Return unmodifiable lists instead of iterators.
The aim was to encapsulate the lists, but iterators actually don’t do a good job
at that, since they have a remove() method (why!). Would be nice if there was a
List-like interface that did not expose any modification methods, but alas.
Second best option is Collections.unmodifiableList(), which restricts the list
access using runtime exceptions.