Message: Fix data byte out of range check.
It was comparing against 1 while it should have been comparing against 0x80.
It allows 0xF7 (EOX) as last byte when the status code is 0xF0 (sysex).

It’s a bit of a layering violation, but unfortunately necessary.
Tbh I don’t think EOX should be included in the data, but that’s what Java does.