pom: Upgrade to OS X app bundler 1.0 alpha 3 for Xcode 4.3 compatibility.
This fixes the issue where the /Developer folder disappeared and you had
to symlink it. The corresponding note is removed from the README.md.

Because for some reason this version isn’t available on maven-central yet,
I’ve uploaded it to a repo on grauw.nl. Because this is a simple plain-file
repository containing only this plugin, only use it as a fallback to
maven-central so it doesn’t get spammed with 404s.

You can also mvn install osxappbundle-maven-plugin locally, from:
http://svn.codehaus.org/mojo/tags/osxappbundle-maven-plugin-1.0-alpha-3
http://svn.codehaus.org/mojo/trunk/mojo/osxappbundle-maven-plugin